pyproject.toml
[project]
name = "sprint-manager-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Bhargav Tumati",email = "bhargav.tumati@symphonize.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
